Storybook announce + AOE Mobskills & Autopot & Inventory check Patch
Implemented storybook detection for the skillbook announcer.
Fixed an issue with AOE mob skills not locating players properly in certain situations.
Reviewed interaction of players within same remote network on multiclient system.
Patched messages when leaving the expedition area showing up for leaving players.
Patched some skill questline item chances unexpectedly too low.
Reviewed Rush questline-exclusive area not allowing multiple players in the room.
Fixed meso capacity check on shop owner happening after item transaction.
Patched server-side autopot settings, now using threshold cooling method to determine a player's updated settings.
Reviewed locking flow regarding world/channel deployment on main class.
Fixed a scenario where items collectable by party members would not show up in case those were the only items to be dropped and none needed by the last-hitter.
Patched an issue in the inventory space checking algorithm that would allow a transaction for multiple of the same equipments.
Refactored Dojo entry/exit modules.
Patched a check on Maker skill.
Fixed Cygnus FA buff getting reapplied each time the skill procs.

+++ engine/engine-base/src/main/java/server/processor/StatEffectProcessor.java
@@ -403,8 +403,8 @@ public class StatEffectProcessor {
       ret.setDamage(MapleDataTool.getIntConvert("damage", source, 100));
       ret.setFixdamage(MapleDataTool.getIntConvert("fixdamage", source, -1));
       ret.setAttackCount(MapleDataTool.getIntConvert("attackCount", source, 1));
-      ret.setBulletCount((byte) MapleDataTool.getIntConvert("bulletCount", source, 1));
-      ret.setBulletConsume((byte) MapleDataTool.getIntConvert("bulletConsume", source, 0));
+      ret.setBulletCount((short) MapleDataTool.getIntConvert("bulletCount", source, 1));
+      ret.setBulletConsume((short) MapleDataTool.getIntConvert("bulletConsume", source, 0));
       ret.setMoneyCon(MapleDataTool.getIntConvert("moneyCon", source, 0));
       ret.setItemCon(MapleDataTool.getInt("itemCon", source, 0));
       ret.setItemConNo(MapleDataTool.getInt("itemConNo", source, 0));
